// lib/Api/api.ts
export interface Article {
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    slug: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    body: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Article
     */
    tagList: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    createdAt: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    updatedAt: string;
    /**
     * 
     * @type {boolean}
     * @memberof Article
     */
    favorited: boolean;
    /**
     * 
     * @type {number}
     * @memberof Article
     */
    favoritesCount: number;
    /**
     * 
     * @type {Profile}
     * @memberof Article
     */
    author: Profile;
}
// ...
export interface NewArticle {
    /**
     * 
     * @type {string}
     * @memberof NewArticle
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof NewArticle
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof NewArticle
     */
    body: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof NewArticle
     */
    tagList?: Array<string>;
}
/**
 * 
 * @export
 * @interface NewArticleRequest
 */
export interface NewArticleRequest {
    /**
     * 
     * @type {NewArticle}
     * @memberof NewArticleRequest
     */
    article: NewArticle;
}
// ...
export interface Profile {
    /**
     * 
     * @type {string}
     * @memberof Profile
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof Profile
     */
    bio: string;
    /**
     * 
     * @type {string}
     * @memberof Profile
     */
    image: string;
    /**
     * 
     * @type {boolean}
     * @memberof Profile
     */
    following: boolean;
}
// ...
export interface UpdateArticle {
    /**
     * 
     * @type {string}
     * @memberof UpdateArticle
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateArticle
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateArticle
     */
    body?: string;
}
/**
 * 
 * @export
 * @interface UpdateArticleRequest
 */
export interface UpdateArticleRequest {
    /**
     * 
     * @type {UpdateArticle}
     * @memberof UpdateArticleRequest
     */
    article: UpdateArticle;
}
// ...
export class ArticlesApi extends BaseAPI {
    /**
     * Create an article. Auth is required
     * @summary Create an article
     * @param {NewArticleRequest} article Article to create
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ArticlesApi
     */
    public createArticle(article: NewArticleRequest, options?: any) {
        return ArticlesApiFp(this.configuration).createArticle(article, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Delete an article. Auth is required
     * @summary Delete an article
     * @param {string} slug Slug of the article to delete
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ArticlesApi
     */
    public deleteArticle(slug: string, options?: any) {
        return ArticlesApiFp(this.configuration).deleteArticle(slug, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Get an article. Auth not required
     * @summary Get an article
     * @param {string} slug Slug of the article to get
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ArticlesApi
     */
    public getArticle(slug: string, options?: any) {
        return ArticlesApiFp(this.configuration).getArticle(slug, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Get most recent articles globally. Use query parameters to filter results. Auth is optional
     * @summary Get recent articles globally
     * @param {string} [tag] Filter by tag
     * @param {string} [author] Filter by author (username)
     * @param {string} [favorited] Filter by favorites of a user (username)
     * @param {number} [limit] Limit number of articles returned (default is 20)
     * @param {number} [offset] Offset/skip number of articles (default is 0)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ArticlesApi
     */
    public getArticles(tag?: string, author?: string, favorited?: string, limit?: number, offset?: number, options?: any) {
        return ArticlesApiFp(this.configuration).getArticles(tag, author, favorited, limit, offset, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Get most recent articles from users you follow. Use query parameters to limit. Auth is required
     * @summary Get recent articles from users you follow
     * @param {number} [limit] Limit number of articles returned (default is 20)
     * @param {number} [offset] Offset/skip number of articles (default is 0)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ArticlesApi
     */
    public getArticlesFeed(limit?: number, offset?: number, options?: any) {
        return ArticlesApiFp(this.configuration).getArticlesFeed(limit, offset, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Update an article. Auth is required
     * @summary Update an article
     * @param {string} slug Slug of the article to update
     * @param {UpdateArticleRequest} article Article to update
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ArticlesApi
     */
    public updateArticle(slug: string, article: UpdateArticleRequest, options?: any) {
        return ArticlesApiFp(this.configuration).updateArticle(slug, article, options).then((request) => request(this.axios, this.basePath));
    }
}